Migrate exp tests
