+++ src/unittests/f32_interval.spec.ts
@@ -24,7 +24,6 @@ import {
   distanceInterval,
   divisionInterval,
   dotInterval,
-  expInterval,
   exp2Interval,
   faceForwardIntervals,
   floorInterval,
@@ -1920,32 +1919,6 @@ interface ScalarToIntervalCase {
   input: number;
   expected: number | IntervalBounds;
 }
-
-g.test('expInterval')
-  .paramsSubcasesOnly<ScalarToIntervalCase>(
-    // prettier-ignore
-    [
-      { input: kValue.f32.infinity.negative, expected: kAnyBounds },
-      { input: 0, expected: 1 },
-      { input: 1, expected: [kValue.f32.positive.e, plusOneULP(kValue.f32.positive.e)] },
-      { input: 89, expected: kAnyBounds },
-    ]
-  )
-  .fn(t => {
-    const error = (x: number): number => {
-      const n = 3 + 2 * Math.abs(t.params.input);
-      return n * oneULPF32(x);
-    };
-
-    t.params.expected = applyError(t.params.expected, error);
-    const expected = toF32Interval(t.params.expected);
-
-    const got = expInterval(t.params.input);
-    t.expect(
-      objectEquals(expected, got),
-      `expInterval(${t.params.input}) returned ${got}. Expected ${expected}`
-    );
-  });
 
 g.test('exp2Interval')
   .paramsSubcasesOnly<ScalarToIntervalCase>(
