+++ src/unittests/floating_point.spec.ts
@@ -457,3 +457,29 @@ g.test('degreesInterval_f32')
       `f32.degreesInterval(${t.params.input}) returned ${got}. Expected ${expected}`
     );
   });
+
+g.test('expInterval_f32')
+  .paramsSubcasesOnly<ScalarToIntervalCase>(
+    // prettier-ignore
+    [
+      { input: kValue.f32.infinity.negative, expected: kAnyBounds },
+      { input: 0, expected: 1 },
+      { input: 1, expected: [kValue.f32.positive.e, plusOneULPF32(kValue.f32.positive.e)] },
+      { input: 89, expected: kAnyBounds },
+    ]
+  )
+  .fn(t => {
+    const error = (x: number): number => {
+      const n = 3 + 2 * Math.abs(t.params.input);
+      return n * oneULPF32(x);
+    };
+
+    t.params.expected = applyError(t.params.expected, error);
+    const expected = FP.f32.toInterval(t.params.expected);
+
+    const got = FP.f32.expInterval(t.params.input);
+    t.expect(
+      objectEquals(expected, got),
+      `f32.expInterval(${t.params.input}) returned ${got}. Expected ${expected}`
+    );
+  });
